gpu blob filter cleanup (#404)
* Clean up gpu blob filter
* use more appropriate name
* remove unused
* fix dlib cmake logic
* simplify ios build script
* fix narrowing errors; cmake typo
note: minor utility function duplication

+++ src/app/dlib/eyexml.cpp
@@ -84,7 +84,7 @@ static cv::Size read_png_size(const std::string &filename)
         in.seekg(16);
         in.read((char *)&width, 4);
         in.read((char *)&height, 4);
-        size = { ntoh_any(width), ntoh_any(height) };
+        size = { static_cast<int>(ntoh_any(width)), static_cast<int>(ntoh_any(height)) };
     }
     return size;
 }
